Make some corrections

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,17 +10,17 @@ int main(void)
 {
 	int i;
 
-	unsigned long double sum = 1.0;
-	unsigned long double j = 2.0;
-	unsigned long double k;
+	unsigned long float sum = 1.0;
+	unsigned long float j = 2.0;
+	unsigned long float k;
 
-	printf("%.0Lf, ", sum);
+	printf("%.0lf, ", sum);
 	for (i = 1; i < 98; i++)
 	{
 		k = sum + j; /* at first run, sum =3*/
 		sum = j;
 		j = k;
-		printf("%.0Lf", sum); /* use capital L for long double*/
+		printf("%.0lf", sum); /* use capital L for long double*/
 
 		if (i != 97)
 		{
